refactor: divide services lifecycle to a separated class
